+++ src/BookFx/PageMargins.cs
@@ -32,7 +32,7 @@
         /// Creates page margins in centimetres with the same top, right, bottom and left margins.
         /// </summary>
         [Pure]
-        public static PageMargins InCentimetres(decimal margin) =>
+        public static PageMargins InCentimetres(double margin) =>
             PageMarginsCore.InCentimetres.With(top: margin, right: margin, bottom: margin, left: margin);
 
         /// <summary>
@@ -42,7 +42,7 @@
         /// <param name="vertical">Top and bottom margin.</param>
         /// <param name="horizontal">Left and right margin.</param>
         [Pure]
-        public static PageMargins InCentimetres(decimal vertical, decimal horizontal) =>
+        public static PageMargins InCentimetres(double vertical, double horizontal) =>
             PageMarginsCore.InCentimetres.With(top: vertical, right: horizontal, bottom: vertical, left: horizontal);
 
         /// <summary>
@@ -53,7 +53,7 @@
         /// <param name="bottom">A bottom margin.</param>
         /// <param name="left">A left margin.</param>
         [Pure]
-        public static PageMargins InCentimetres(decimal top, decimal right, decimal bottom, decimal left) =>
+        public static PageMargins InCentimetres(double top, double right, double bottom, double left) =>
             PageMarginsCore.InCentimetres.With(top: top, right: right, bottom: bottom, left: left);
 
         /// <summary>
@@ -66,7 +66,7 @@
         /// Creates page margins in inches with the same top, right, bottom and left margins.
         /// </summary>
         [Pure]
-        public static PageMargins InInches(decimal margin) =>
+        public static PageMargins InInches(double margin) =>
             PageMarginsCore.InInches.With(top: margin, right: margin, bottom: margin, left: margin);
 
         /// <summary>
@@ -76,7 +76,7 @@
         /// <param name="vertical">Top and bottom margin.</param>
         /// <param name="horizontal">Left and right margin.</param>
         [Pure]
-        public static PageMargins InInches(decimal vertical, decimal horizontal) =>
+        public static PageMargins InInches(double vertical, double horizontal) =>
             PageMarginsCore.InInches.With(top: vertical, right: horizontal, bottom: vertical, left: horizontal);
 
         /// <summary>
@@ -87,43 +87,43 @@
         /// <param name="bottom">A bottom margin.</param>
         /// <param name="left">A left margin.</param>
         [Pure]
-        public static PageMargins InInches(decimal top, decimal right, decimal bottom, decimal left) =>
+        public static PageMargins InInches(double top, double right, double bottom, double left) =>
             PageMarginsCore.InInches.With(top: top, right: right, bottom: bottom, left: left);
 
         /// <summary>
         /// Define a top margin.
         /// </summary>
         [Pure]
-        public PageMargins Top(decimal margin) => Get.With(top: margin);
+        public PageMargins Top(double margin) => Get.With(top: margin);
 
         /// <summary>
         /// Define a right margin.
         /// </summary>
         [Pure]
-        public PageMargins Right(decimal margin) => Get.With(right: margin);
+        public PageMargins Right(double margin) => Get.With(right: margin);
 
         /// <summary>
         /// Define a bottom margin.
         /// </summary>
         [Pure]
-        public PageMargins Bottom(decimal margin) => Get.With(bottom: margin);
+        public PageMargins Bottom(double margin) => Get.With(bottom: margin);
 
         /// <summary>
         /// Define a left margin.
         /// </summary>
         [Pure]
-        public PageMargins Left(decimal margin) => Get.With(left: margin);
+        public PageMargins Left(double margin) => Get.With(left: margin);
 
         /// <summary>
         /// Define a header margin.
         /// </summary>
         [Pure]
-        public PageMargins Header(decimal margin) => Get.With(header: margin);
+        public PageMargins Header(double margin) => Get.With(header: margin);
 
         /// <summary>
         /// Define a footer margin.
         /// </summary>
         [Pure]
-        public PageMargins Footer(decimal margin) => Get.With(footer: margin);
+        public PageMargins Footer(double margin) => Get.With(footer: margin);
     }
 }
